requirements.txt
colorama==0.4.4
numpy==1.19.5
Pillow==8.4.0
